Move fetch wrapper into utils file
