+++ src/utils.js
@@ -1,5 +1,11 @@
 import pluralize from 'pluralize';
 
+export const jsonContentTypes = [
+  'application/json',
+  'application/vnd.api+json'
+];
+
+export const noop = () => {};
 export const copyState = state => JSON.parse(JSON.stringify(state));
 
 export const findEntity = (state, { type, id }) => {
@@ -31,4 +37,29 @@ export const findForeignKeyInEntity = (entity, foreignKeyType) => {
   });
 
   return foreignKey;
+};
+
+export const apiRequest = (url, accessToken, options = {}) => {
+  const allOptions = {
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/vnd.api+json'
+    },
+    ...options
+  };
+
+  return fetch(url, allOptions)
+    .then(res => {
+      if (res.status >= 200 && res.status < 300) {
+        if (jsonContentTypes.indexOf(res.headers.get('Content-Type')) > -1) {
+          return res.json();
+        }
+
+        return res;
+      }
+
+      const e = new Error(res.statusText);
+      e.response = res;
+      throw e;
+    });
 };
